fix: corregida configuración y entidades para el arranque correcto - añadida dependencia flyway-database-postgresql en pom.xml - corregidas anotaciones JPA en ClienteDatos, ClientePrivado, ResultadoAnalisis y Usuario - añadidas propiedades jwt.secret, jwt.expiration y clustering.service.url - corregidos tipos NUMERIC a DOUBLE PRECISION en migraciones V3 y V4 - corregido hash BCrypt para admin1234 en V5 - añadido bean RestTemplate en AppConfig y Dockerfile del backend - ajustadas variables en docker-compose.yml y docker-compose.dev.yml

+++ backend/ChurnGym/src/main/resources/db/migration/V3__create_clientes_datos.sql
@@ -8,7 +8,7 @@ CREATE TABLE clientes_datos (
     meses_como_socio    INTEGER,
     franja_horaria      VARCHAR(10)   NOT NULL CHECK (franja_horaria IN ('MANANA', 'TARDE', 'NOCHE')),
     deporte_principal   VARCHAR(20)   NOT NULL CHECK (deporte_principal IN ('MUSCULACION', 'CARDIO', 'CROSSFIT', 'CLASES_COLECTIVAS', 'MIXTO')),
-    frecuencia_semanal  NUMERIC(4,1),
+    frecuencia_semanal  DOUBLE PRECISION,
     semanas_inactivo    INTEGER,
-    tendencia_mensual   NUMERIC(5,1)
+    tendencia_mensual   DOUBLE PRECISION
 );
